Added auction card
