+++ src/components/Auctions.js
@@ -1,6 +1,22 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
+
+import AuctionCard from './AuctionCard';
+
+import { axiosWithAuth } from '../utils/axiosWithAuth';
 
 const Auctions = (props) => {
+  const [data, setData] = useState([]);
+
+  useEffect(() => {
+    axiosWithAuth().get('/api/auctions')
+      .then(res => {
+        setData(res.data);
+      })
+      .catch(err => {
+        console.log(err);
+      });
+  }, [])
+  console.log(data);
   const auctions = [{
     name: "Iphone"
   },
@@ -10,11 +26,10 @@ const Auctions = (props) => {
 ]
   return (
     <div>
-      {auctions.map(auction => 
-        <div>
-          {auction.name}
-        </div>
-          )}
+      hi
+      {data.map(auction => 
+        <AuctionCard auction={auction} />  
+      )}
     </div>
   )
 }
